remove miniflare test worker (not needed with node 18+)
