--- packages/api/test/test-worker/wrangler.toml
@@ -1,23 +1,0 @@
-# Development
-
-name = "nft-storage-test-worker"
-main = "./dist/worker.js"
-account_id = "fffa4b4363a7e5250af8357087263b3a"
-workers_dev = true
-
-# Compatibility flags https://github.com/cloudflare/wrangler/pull/2009
-compatibility_date = "2021-08-23"
-compatibility_flags = [ "formdata_parser_supports_files" ]
-
-[vars]
-ENV = "test"
-DEBUG = "true"
-DATABASE_URL = "http://localhost:3000"
-
-[build]
-command = "scripts/build.js"
-
-# route definitions 
-[miniflare]
-route = "http://127.0.0.1/test-worker*"
-routes = ["test.mf/*"]
